+++ src/main/java/xyz/nikitacartes/easyauth/config/LangConfigV1.java
@@ -1,5 +1,7 @@
 package xyz.nikitacartes.easyauth.config;
 
+import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
 import net.minecraft.server.command.CommandOutput;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.server.network.ServerPlayerEntity;
@@ -7,9 +9,17 @@ import net.minecraft.text.MutableText;
 import net.minecraft.text.Text;
 import org.spongepowered.configurate.objectmapping.ConfigSerializable;
 import org.spongepowered.configurate.objectmapping.meta.Comment;
+import xyz.nikitacartes.easyauth.utils.EasyLogger;
+
+import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.lang.reflect.Type;
+import java.util.HashMap;
+import java.util.Map;
 
 import static net.minecraft.text.Text.translatable;
 import static net.minecraft.text.Text.translatableWithFallback;
+import static net.minecraft.text.TranslatableTextContent.EMPTY_ARGUMENTS;
 import static xyz.nikitacartes.easyauth.EasyAuth.langConfig;
 
 @ConfigSerializable
@@ -17,51 +27,61 @@ public class LangConfigV1 extends ConfigTemplate {
 
     @Comment("""
             Enable server-side translation.
-            While enabaled EasyAuth sends messages, translated to player's client language.
-            List of aviailable languages: https://github.com/NikitaCartes/EasyAuth/tree/HEAD/src/main/resources/data/easyauth/lang
-            Disabling this oprion will force EasyAuth to send all messaged from that file.""")
+            While enabled EasyAuth sends messages, translated to player's client language.
+            List of available languages: https://github.com/NikitaCartes/EasyAuth/tree/HEAD/src/main/resources/data/easyauth/lang
+            Disabling this option will force EasyAuth to send all messaged from that file.""")
     public boolean enableServerSideTranslation = true;
-    public TranslatableText enterPassword = new TranslatableText("text.easyauth.enterPassword", "§6You need to enter your password!");
-    public TranslatableText enterNewPassword = new TranslatableText("text.easyauth.enterNewPassword", "§4You need to enter new password!");
-    public TranslatableText wrongPassword = new TranslatableText("text.easyauth.wrongPassword", "§4Wrong password!");
-    public TranslatableText matchPassword = new TranslatableText("text.easyauth.matchPassword", "§6Passwords must match!");
-    public TranslatableText passwordUpdated = new TranslatableText("text.easyauth.passwordUpdated", "§aYour password was updated successfully!");
-    public TranslatableText loginRequired = new TranslatableText("text.easyauth.loginRequired", "§cYou are not authenticated!\n§6Use /login, /l to authenticate!");
-    public TranslatableText loginTriesExceeded = new TranslatableText("text.easyauth.loginTriesExceeded", "§4Too many login tries. Please wait a few minutes and try again.");
-    public TranslatableText globalPasswordSet = new TranslatableText("text.easyauth.globalPasswordSet", "§aGlobal password was successfully set!");
-    public TranslatableText cannotChangePassword = new TranslatableText("text.easyauth.cannotChangePassword", "§cYou cannot change password!");
-    public TranslatableText cannotUnregister = new TranslatableText("text.easyauth.cannotUnregister", "§cYou cannot unregister this account!");
-    public TranslatableText notAuthenticated = new TranslatableText("text.easyauth.notAuthenticated", "§cYou are not authenticated!\n§6Try with /login, /l or /register.");
-    public TranslatableText alreadyAuthenticated = new TranslatableText("text.easyauth.alreadyAuthenticated", "§6You are already authenticated.");
-    public TranslatableText successfullyAuthenticated = new TranslatableText("text.easyauth.successfullyAuthenticated", "§aYou are now authenticated.");
-    public TranslatableText successfulLogout = new TranslatableText("text.easyauth.successfulLogout", "§aLogged out successfully.");
-    public TranslatableText timeExpired = new TranslatableText("text.easyauth.timeExpired", "§cTime for authentication has expired.");
-    public TranslatableText registerRequired = new TranslatableText("text.easyauth.registerRequired", "§6Type /register <password> <password> to claim this account.");
-    public TranslatableText alreadyRegistered = new TranslatableText("text.easyauth.alreadyRegistered", "§6This account name is already registered!");
-    public TranslatableText registerSuccess = new TranslatableText("text.easyauth.registerSuccess", "§aYou are now authenticated.");
-    public TranslatableText userdataDeleted = new TranslatableText("text.easyauth.userdataDeleted", "§aUserdata deleted.");
-    public TranslatableText userdataUpdated = new TranslatableText("text.easyauth.userdataUpdated", "§aUserdata updated.");
-    public TranslatableText accountDeleted = new TranslatableText("text.easyauth.accountDeleted", "§aYour account was successfully deleted!");
-    public TranslatableText configurationReloaded = new TranslatableText("text.easyauth.configurationReloaded", "§aConfiguration file was reloaded successfully.");
-    public TranslatableText maxPasswordChars = new TranslatableText("text.easyauth.maxPasswordChars", "§6Password can be at most %d characters long!");
-    public TranslatableText minPasswordChars = new TranslatableText("text.easyauth.minPasswordChars", "§6Password needs to be at least %d characters long!");
-    public TranslatableText disallowedUsername = new TranslatableText("text.easyauth.disallowedUsername", "§6Invalid username characters! Allowed character regex: %s");
-    public TranslatableText playerAlreadyOnline = new TranslatableText("text.easyauth.playerAlreadyOnline", "§cPlayer %s is already online!");
-    public TranslatableText worldSpawnSet = new TranslatableText("text.easyauth.worldSpawnSet", "§aSpawn for logging in was set successfully.");
-    public TranslatableText corruptedPlayerData = new TranslatableText("text.easyauth.corruptedPlayerData", "§cYour data is probably corrupted. Please contact admin.");
-    public TranslatableText userNotRegistered = new TranslatableText("text.easyauth.userNotRegistered", "§cThis player is not registered!");
-    public TranslatableText cannotLogout = new TranslatableText("text.easyauth.cannotLogout", "§cYou cannot logout!");
-    public TranslatableText offlineUuid = new TranslatableText("text.easyauth.offlineUuid", "Offline UUID for %s is %s");
-    public TranslatableText registeredPlayers = new TranslatableText("text.easyauth.registeredPlayers", "List of registered players:");
-    public TranslatableText validSession = new TranslatableText("text.easyauth.validSession", "§aYou have a valid session. No need to log in.");
-    public TranslatableText onlinePlayerLogin = new TranslatableText("text.easyauth.onlinePlayerLogin", "§aYou are using an online account. No need to log in.");
-    public TranslatableText differentUsernameCase = new TranslatableText("text.easyauth.diffrentUsernameCase", "§6You are using a different case of your username. Please use the correct one.");
-    public TranslatableText wrongGlobalPassword = new TranslatableText("text.easyauth.wrongGlobalPassword", "§4Wrong global password!");
-    public TranslatableText registerRequiredWithGlobalPassword = new TranslatableText("text.easyauth.registerRequiredWithGlobalPassword", "§6Type /register <global password> <password> <password> to claim this account.");
-    public TranslatableText markAsOffline = new TranslatableText("text.easyauth.markAsOffline", "§aPlayer %s was marked as offline.");
-    public TranslatableText markAsOnline = new TranslatableText("text.easyauth.markAsOnline", "§aPlayer %s was marked as online.");
-    public TranslatableText selfMarkAsOnline = new TranslatableText("text.easyauth.selfMarkAsOnline", "§aYou marked yourself as online player. You can rejoin now.");
-    public TranslatableText selfMarkAsOnlineWarning = new TranslatableText("text.easyauth.selfMarkAsOnlineWarning", "§6You want to mark yourself as online player.\n§6You will not be able to log in if you don't have an online account.\n§6Data, connected to offline uuid (villagers' discounts, pets) will be lost.\n§aIf you are want to continue, type /account online <password> true.");
+
+    @Comment("""
+            
+            Default language for EasyAuth.
+            
+            Note: with server-side translation enabled, this language will be used for non-translatable messages
+            Note: with disable server-side translation, message from "text" field have higher priority than defaultLanguage.""")
+    public String defaultLanguage = "en_us";
+    public TranslatableText enterPassword = new TranslatableText();
+    public TranslatableText enterNewPassword = new TranslatableText();
+    public TranslatableText wrongPassword = new TranslatableText();
+    public TranslatableText matchPassword = new TranslatableText();
+    public TranslatableText passwordUpdated = new TranslatableText();
+    public TranslatableText loginRequired = new TranslatableText();
+    public TranslatableText loginTriesExceeded = new TranslatableText();
+    public TranslatableText globalPasswordSet = new TranslatableText();
+    public TranslatableText cannotChangePassword = new TranslatableText();
+    public TranslatableText cannotUnregister = new TranslatableText();
+    public TranslatableText notAuthenticated = new TranslatableText();
+    public TranslatableText alreadyAuthenticated = new TranslatableText();
+    public TranslatableText successfullyAuthenticated = new TranslatableText();
+    public TranslatableText successfulLogout = new TranslatableText();
+    public TranslatableText timeExpired = new TranslatableText();
+    public TranslatableText registerRequired = new TranslatableText();
+    public TranslatableText alreadyRegistered = new TranslatableText();
+    public TranslatableText registerSuccess = new TranslatableText();
+    public TranslatableText userdataDeleted = new TranslatableText();
+    public TranslatableText userdataUpdated = new TranslatableText();
+    public TranslatableText accountDeleted = new TranslatableText();
+    public TranslatableText configurationReloaded = new TranslatableText();
+    public TranslatableText maxPasswordChars = new TranslatableText();
+    public TranslatableText minPasswordChars = new TranslatableText();
+    public TranslatableText disallowedUsername = new TranslatableText();
+    public TranslatableText playerAlreadyOnline = new TranslatableText();
+    public TranslatableText worldSpawnSet = new TranslatableText();
+    public TranslatableText corruptedPlayerData = new TranslatableText();
+    public TranslatableText userNotRegistered = new TranslatableText();
+    public TranslatableText cannotLogout = new TranslatableText();
+    public TranslatableText offlineUuid = new TranslatableText();
+    public TranslatableText registeredPlayers = new TranslatableText();
+    public TranslatableText validSession = new TranslatableText();
+    public TranslatableText onlinePlayerLogin = new TranslatableText();
+    public TranslatableText differentUsernameCase = new TranslatableText();
+    public TranslatableText wrongGlobalPassword = new TranslatableText();
+    public TranslatableText registerRequiredWithGlobalPassword = new TranslatableText();
+    public TranslatableText markAsOffline = new TranslatableText();
+    public TranslatableText markAsOnline = new TranslatableText();
+    public TranslatableText selfMarkAsOnline = new TranslatableText();
+    public TranslatableText selfMarkAsOnlineWarning = new TranslatableText();
+
+    private static Map<String, String> translations = new HashMap<>();
 
     public LangConfigV1() {
         super("translation.conf", """
@@ -76,6 +96,7 @@ public class LangConfigV1 extends ConfigTemplate {
         if (config == null) {
             config = new LangConfigV1();
             config.save();
+            config = loadConfig(LangConfigV1.class, "translation.conf");
         }
         return config;
     }
@@ -85,6 +106,26 @@ public class LangConfigV1 extends ConfigTemplate {
         if (config == null) {
             throw new RuntimeException("Failed to load translation.conf");
         }
+
+        Gson gson = new Gson();
+        Type mapType = new TypeToken<Map<String, String>>() {}.getType();
+
+        String path = "data/easyauth/lang/" + config.defaultLanguage + ".json";
+        InputStream input = LangConfigV1.class.getClassLoader().getResourceAsStream(path);
+
+        if (input == null) {
+            EasyLogger.LogError("Failed to load default language " + config.defaultLanguage + ".json. Using en_us.json instead.");
+            translations = new HashMap<>();
+            return config;
+        }
+
+        try (InputStreamReader reader = new InputStreamReader(input)) {
+            translations = gson.fromJson(reader, mapType);
+        } catch (Exception e) {
+            EasyLogger.LogError("Failed to load default language", e);
+            translations = new HashMap<>();
+        }
+
         return config;
     }
 
@@ -98,6 +139,13 @@ public class LangConfigV1 extends ConfigTemplate {
         public final String fallback;
         public final boolean enabled;
         public final boolean serverSide;
+
+        public TranslatableText() {
+            this.key = null;
+            this.fallback = "";
+            this.enabled = true;
+            this.serverSide = true;
+        }
 
         public TranslatableText(String key, String fallback) {
             this.key = key;
@@ -115,51 +163,31 @@ public class LangConfigV1 extends ConfigTemplate {
 
         public void send(ServerCommandSource commandOutput) {
             if (enabled && commandOutput != null) {
-                if (langConfig.enableServerSideTranslation && serverSide) {
-                    commandOutput.sendMessage(translatable(key));
-                } else {
-                    commandOutput.sendMessage(Text.literal(fallback));
-                }
+                commandOutput.sendMessage(getTranslation());
             }
         }
 
         public void send(ServerPlayerEntity commandOutput) {
             if (enabled && commandOutput != null) {
-                if (langConfig.enableServerSideTranslation && serverSide) {
-                    commandOutput.sendMessage(translatable(key));
-                } else {
-                    commandOutput.sendMessage(Text.literal(fallback));
-                }
+                commandOutput.sendMessage(getTranslation());
             }
         }
 
         public <T extends CommandOutput> void send(T commandOutput) {
             if (enabled && commandOutput != null) {
-                if (langConfig.enableServerSideTranslation && serverSide) {
-                    commandOutput.sendMessage(translatable(key));
-                } else {
-                    commandOutput.sendMessage(Text.literal(fallback));
-                }
+                commandOutput.sendMessage(getTranslation());
             }
         }
 
         public void send(ServerCommandSource commandOutput, Object... args) {
             if (enabled && commandOutput != null) {
-                if (langConfig.enableServerSideTranslation && serverSide) {
-                    commandOutput.sendMessage(translatable(key, args));
-                } else {
-                    commandOutput.sendMessage(translatable(fallback, args));
-                }
+                commandOutput.sendMessage(getTranslation(args));
             }
         }
 
         public MutableText get() {
             if (enabled) {
-                if (langConfig.enableServerSideTranslation && serverSide) {
-                    return translatable(key);
-                } else {
-                    return Text.literal(fallback);
-                }
+                return getTranslation();
             } else {
                 return Text.literal("");
             }
@@ -167,37 +195,39 @@ public class LangConfigV1 extends ConfigTemplate {
 
         public MutableText get(Object... args) {
             if (enabled) {
-                if (langConfig.enableServerSideTranslation && serverSide) {
-                    return translatable(key, args);
-                } else {
-                    return translatable(fallback, args);
-                }
+                return getTranslation(args);
             } else {
                 return Text.literal("");
             }
         }
 
-        public MutableText getWithFallback() {
-            if (enabled) {
-                if (langConfig.enableServerSideTranslation && serverSide) {
-                    return translatableWithFallback(key, fallback);
-                } else {
-                    return Text.literal(fallback);
-                }
-            } else {
-                return Text.literal("");
-            }
+        public MutableText getNonTranslatable() {
+            return getNonTranslatable(EMPTY_ARGUMENTS);
         }
 
-        public MutableText getWithFallback(Object... args) {
+        public MutableText getNonTranslatable(Object... args) {
             if (enabled) {
-                if (langConfig.enableServerSideTranslation && serverSide) {
+                if (!fallback.isEmpty()) {
                     return translatableWithFallback(key, fallback, args);
                 } else {
-                    return translatable(fallback, args);
+                    return translatableWithFallback(key, translations.get(key), args);
                 }
             } else {
                 return Text.literal("");
+            }
+        }
+
+        private MutableText getTranslation() {
+            return getTranslation(EMPTY_ARGUMENTS);
+        }
+
+        private MutableText getTranslation(Object... args) {
+            if (langConfig.enableServerSideTranslation && serverSide) {
+                return translatable(key, args);
+            } if (!fallback.isEmpty()) {
+                return translatableWithFallback(key, fallback, args);
+            } else {
+                return translatableWithFallback(key, translations.get(key), args);
             }
         }
     }